renamed, made internal
